# Token become immutable, performance optimizations