--- a/phpunit.xml.dist
+++ b/phpunit.xml.dist
@@ -43,5 +43,6 @@
     <php>
         <ini name="zend.enable_gc" value="0"/>
         <env name="SKIP_LINT_TEST_CASES" value="0"/>
+        <env name="PHP_CS_FIXER_TEST_USE_LEGACY_TOKENIZER" value="0"/>
     </php>
 </phpunit>