--- a/src/Fixer/Import/NoLeadingImportSlashFixer.php
+++ b/src/Fixer/Import/NoLeadingImportSlashFixer.php
@@ -56,12 +56,12 @@ final class NoLeadingImportSlashFixer extends AbstractFixer
      */
     protected function applyFix(\SplFileInfo $file, Tokens $tokens)
     {
-        $foundNamespace = $tokens->findGivenKind(T_NAMESPACE);
-        if (empty($foundNamespace)) {
+        if (!$tokens->isTokenKindFound(T_NAMESPACE)) {
             return;
         }
 
         $tokensAnalyzer = new TokensAnalyzer($tokens);
+        $foundNamespace = $tokens->findGivenKind(T_NAMESPACE);
         $firstNamespaceIdx = key($foundNamespace);
 
         $usesIdxs = $tokensAnalyzer->getImportUseIndexes();