--- a/src/Runner/Runner.php
+++ b/src/Runner/Runner.php
@@ -164,6 +164,8 @@ final class Runner
         }
 
         $old = file_get_contents($file->getRealPath());
+
+        Tokens::setLegacyMode(false);
         $tokens = Tokens::fromCode($old);
         $oldHash = $tokens->getCodeHash();