TODO:
- [x] rebase on master

```
Executing on Symfony 2.7 with 45files that reports violations

before
> Checked all files in 77.743 seconds, 20.000 MB memory used
> Checked all files in 78.761 seconds, 20.000 MB memory used

after improvement
> Checked all files in 70.417 seconds, 20.000 MB memory used
> Checked all files in 71.144 seconds, 20.000 MB memory used

About 10% improvement
```

Deps:
- [x] #2819
- [x] #2816
- [x] #2817
- [x] #2823